Reading from file finished. I have prepare some methods in 'Graph' class for you.
addNode(node) - something to store this psi factors inside graph. You have to initialize new object of class 'Node', fill values accordingly and use the function.
delNode(nameOrAlias) - just deletes factor from graph.
getFactors(nameOrAlias) - returns a list of factors, that contains nodes, which are either the node with the name or alias specified or nodes, that have specified node as parents.
getPrVal(nodename, assignments) - this returns a value, which corresponds to probability stored in node of name of alias 'nodename' and with given assignment of variables. The assignment is a list of values for each variable. First value is the value for the node 'nodename' and the rest is the assignment for every parent in order of appearing on the list 'parents' of the node nodename.

Last two prints in bn_inference.py are examples of usage for last two functions.